Correct wiring of Main to Detail View Controller
The navigation controller Identity > Storyboard ID was set to DetailNavigation. This was incorrect. I moved it to Detail View Controller > Identity > Storyboard ID. This is why the custom class wasn't matching up. Because (by ID) I was trying to load the Nav Controller and not the Detail VC. Why did we have the NavC here again? Was it to give us WYSIWYG controls in the Detail VC because we put it into its own file?