Fixed crash bug. Removed didSet
didSet wasn't necessary because updateViews() is being called other places and getting the job done.
Because didSet was getting called in the segue during seutp the view wasn't around yet. Interesting this was handled for the text fields with a ? but the datePicker was not and was causing the crash,