Customer camera / gif maker
Hello @dimitribouniol here is our proposal for an app we would like to create. The main theme of the app is to have a custom camera while will let us control the way a image is taken. While allowing us to create gifs out of the images. 
Please also follow this link to see our current wireframe 
https://whimsical.com/photoscape-9DhxBQJXLwj6LT7U8p8mL4
@maybemichael